requirements.txt
django==1.8
mySQL-python==1.2.5
djangorestframework==3.1
markdown==2.6.2
django-filter==0.10.0